Improved runtime for the main JS function
Reduced the lines of code from 12 all the way to 6, which would help improve the runtime for our project. I used the built in Math object to calculate the value of a base raised to the power of an exponent. Earlier I used a for loop for exponents which had O(n) runtime, where n was the age difference between current and desired age. Now, this has been reduced to O(1) runtime. I also got rid of unnecessary constants.