Final Algorithm for CS222 project
This is the final version of the main algorithm upon which our project is based on. This algorithm does exactly what we intended to do, and is the exact codified version of the formula that was submitted by us in the Project Proposal.

I made 5 realistic test cases to test out the algorithm, and I am pleased to say that the output of all test cases matched the output of similar Retirement calculator apps out there available on the App store. This means that the Algorithm is now fully functional and can be fully integrated to be used as the back-end function for our website. Code is in JS (Javascript).